auth-service/requirements.txt
# auth-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-keycloak==3.7.0
python-jose[cryptography]==3.3.0
httpx==0.25.0
structlog==23.2.0
pydantic==2.4.2
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-exporter-otlp-proto-grpc==1.21.0
python-multipart==0.0.6